requirements.txt
selenium==4.15.0
beautifulsoup4==4.12.2
pandas==2.1.3
numpy==1.24.3
requests==2.31.0
lightgbm==4.1.0
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
googlemaps==4.10.0
geopy==2.4.0
shap==0.43.0
jupyter==1.0.0
tqdm==4.66.1
fake-useragent==1.4.0
python-dotenv==1.0.0
openpyxl==3.1.2